refactor: keep two main layouts – root layout and pages layout with grid
