--- app/[lang]/[...unknown-route]/not-found.tsx
@@ -1,5 +1,0 @@
-import { PageNotFound } from './page-not-found/PageNotFound';
-
-export default function CustomNotFoundPage() {
-  return <PageNotFound />;
-}
