+++ app/[lang]/not-found.tsx
@@ -1,0 +1,5 @@
+import { PageNotFound } from './[...unknown-route]/page-not-found/PageNotFound';
+
+export default function CustomNotFoundPage() {
+  return <PageNotFound />;
+}
